+++ lua/plugins/configs/aerial.lua
@@ -1,0 +1,13 @@
+local options = {
+    layout = {
+        default_direction = "right",
+        placement = "edge",
+        min_width = 25,
+        max_width = { 40, 0.3 },
+    },
+
+    attach_mode = "global",
+    show_guides = true,
+}
+
+return options